requirements.txt
jupyter
scipy
torchvision
tqdm
torch
opencv-python
imantics
numpy
rasterio
matplotlib
scikit-image
Shapely
pandas
ipdb
Pillow
pytorch_lightning
PyYAML
tensorboardX
